pyproject.toml
[tool.poetry]
name = "notes-django-backend"
version = "0.1.0"
description = "Backend for Notes Android Apps"
authors = ["perymerdeka <tkjasikk@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
